Move`VllmConfig` from `config/__init__.py` to `config/vllm.py` (#25271)
