+++ vllm/model_executor/layers/quantization/deepspeedfp.py
@@ -9,9 +9,8 @@ import torch.nn.functional as F
 from packaging import version
 
 from vllm.model_executor.layers.linear import LinearBase, LinearMethodBase
-from vllm.model_executor.layers.quantization import QuantizationMethods
-from vllm.model_executor.layers.quantization.base_config import (
-    QuantizationConfig)
+from vllm.model_executor.layers.quantization import (QuantizationConfig,
+                                                     QuantizationMethods)
 from vllm.model_executor.utils import set_weight_attrs
